Fixes #25619 - refactor logging middleware
