+++ config/logging.yaml
@@ -56,7 +56,7 @@
 #     [%L] line number where the logging request was issued
 #     [%M] method name where the logging request was issued
 #     [%X{request}] request-ID set in HTTP headers, or random UUID
-#     [%X{session}] random ID assigned per session
+#     [%X{session}] random ID assigned per session (not available for the initial log line by Rails)
 #     [%X{user_login}] currently logged user if present
 #     [%X{org_id}] current organization id if available
 #     [%X{loc_id}] current location id if available
